+++ examples/advanced.php
@@ -7,8 +7,8 @@ use srag\DataTable\Component\Data\Row\RowData;
 use srag\DataTable\Component\Format\Format;
 use srag\DataTable\Component\Settings\Settings;
 use srag\DataTable\Component\Settings\Sort\SortField;
-use srag\DataTable\Implementation\Column\Action\AbstractActionColumn;
 use srag\DataTable\Implementation\Column\Column;
+use srag\DataTable\Implementation\Column\Formatter\AbstractActionsFormatter;
 use srag\DataTable\Implementation\Column\Formatter\DefaultFormatter;
 use srag\DataTable\Implementation\Column\Formatter\LinkColumnFormatter;
 use srag\DataTable\Implementation\Data\Data;
@@ -36,7 +36,7 @@ function advanced() : string
         ((new Column($DIC, "title", "Title")))->withFormatter(new LinkColumnFormatter($DIC))->withDefaultSort(true),
         (new Column($DIC, "type", "Type"))->withFormatter(new AdvancedExampleFormatter($DIC)),
         (new Column($DIC, "description", "Description"))->withDefaultSelected(false)->withSortable(false),
-        new AdvancedExampleActionColumns($DIC, "actions", "Actions")
+        (new Column($DIC, "actions", "Actions"))->withFormatter(new AdvancedExampleActionsFormatter($DIC))
     ], new AdvancedExampleDataFetcher($DIC)
     ))->withFilterFields([
         "title" => $DIC->ui()->factory()->input()->field()->text("Title"),
@@ -96,11 +96,11 @@ class AdvancedExampleFormatter extends DefaultFormatter
 }
 
 /**
- * Class AdvancedExampleActionColumns
+ * Class AdvancedExampleActionsFormatter
  *
  * @author  studer + raimann ag - Team Custom 1 <support-custom1@studer-raimann.ch>
  */
-class AdvancedExampleActionColumns extends AbstractActionColumn
+class AdvancedExampleActionsFormatter extends AbstractActionsFormatter
 {
 
     /**